requirements.txt
flask
streamlit
sentence-transformers
faiss-cpu
googletrans==4.0.0-rc1
python-dotenv
requests
tqdm